Refactor multiverse rake MethodDefinitionInTask
